break-out tic-tac-toe validators into their own class
